+++ .metals/metals.lock.db
@@ -1,0 +1,6 @@
+#FileLock
+#Tue Jul 09 10:33:03 IST 2019
+hostName=gadgil-Vostro-470
+id=16bd51d1be53d663a50e42f324cb5bb2cdffffd2136
+method=file
+server=10.134.10.33\:41863
