--- partex/server/resources/mydoc.tex
@@ -1,108 +1,0 @@
-\documentclass[master.tex]{subfiles}
-
-\input{preamble}
-\author{Ankit Jaiswal}
-\newcommand{\foo}{foobar}
-\newcommand{\jc}{John \foo Cena}
-\newcommand{\name}[2]{My first name is #1 and second name is #2}
-\newcommand{\withDefault}[2][books]{my friends are #1 and #2}
-
-\newtheorem{defn*}{Definition}[section]
-\newtheorem{theorem}{Theorem}[section]
-\newtheorem{col}[theorem]{Colrollary}
-\newtheorem{lem}[theorem]{Lemma}
-\newtheorem{notes}[defn*]{Notes}
-\newtheorem{aloo}{Aloo}
-%%%%%%%%%%%%%% BEGIN CONTENT: %%%%%%%%%%%%%%
-
-\begin{document}
-  \title{How to Structure a LaTeX Document}
-  \date{December 2004}
-  \subjclass[2010]{03B15 (primary), 20F12, 20F65 (secondary)}
-  \begin{abstract}
-  Parsing can be fun and this page exactly demostrate that.
-  But hey, there is no output without input. That's why I am writing this
-  down for it to behave as an abstarct, thus an input :)
-  \end{abstract}
-
-  \maketitle
-
-  \section{Testing}
-  This follows from \jc the second part of the \textit{remark} above.
-  % parsing comments
-  \begin{defn*}
-    test and testing.
-  \end{defn*}
-  \begin{aloo}
-  test and test
-  \end{aloo}
-  \withDefault{Abhijeet}
-  Now for some remakrs \% about \foo centralizers. %one more comment.
-  \vspace{1cm}
-  \begin{rmk*}[1.1.3]
-    This is a example of nested environment. \name{Ankit}{Jaiswal} \\
-    Also test the line break token. \withDefault[Bhavna]{Abhijeet}
-    \begin{enumerate}
-    \item\label{pehla} If $A \subgroup G$, \medskip then $A$ is abelian if and only if $A \subset
-      C_G(A)$.
-    \item Furthermore, if $A \subgroup Z(G)$, then $A \normsubgroup G$. \newline
-      This follows using \emph{basic} commutativity arguments, which are as follows:
-      \begin{itemize}
-        % testing nested list
-        \item First One.
-        % testing comments.
-        \item Second One.
-      \end{itemize}
-    \end{enumerate}
-  \end{rmk*}
-  \begin{defn*}[definition of the page]
-      For $A \subset G$, we set $N_G(A) := \{g \in G | gAg^{-1} = A\}$ and
-      $C_G(A) := \{g \in G | gag^{-1} = a, \forall a \in A\}$.
-  \end{defn*}
-  Note that $C_G(A) \subset N_G(A)$ and $Z(G) = C_G(G)$.
-
-
-  \section{math section}
-  \begin{equation}
-  if x = 2 - x then 2x = 2, hence x = 1
-  \end{equation}
-
-  {\tiny\begin{theorem}
-  If the \emph{tiny} is followed immediately by begin this fails.
-  \end{theorem}
-  }
-
-  \subsection{Tyring Floats}
-  This can parse graphics as well and label like in \ref{pehla}.
-  \begin{figure}[h]
-      \centering
-      \includegraphics{aloo.jpg}
-      \caption{FigCaption}
-      \label{fig:my_label}
-  \end{figure}
-
-  \subsubsection{specific case: Tables}
-  Trying tables
-  \begin{table}[]
-      \centering
-      \begin{tabular}{c|c}
-          1 & 2 & 3 \\
-          4 & 5 & 6
-      \end{tabular}
-      \caption{TabCaption}
-      \label{tab:my_label}
-  \end{table}
-
-  Lets check some code blocks. I know this wont be look good. But real document
-  do have a lot plain text material which gives a nice visual appeal.
-
-\section{Tyring code blocks}
-  \begin{lstlisting}
-  void main() {
-    printf("Hello, World!\n");
-  }
-  \end{lstlisting}
-
-  \subsection{specific case of code style}
-  This is to test the latex Numbering.
-\end{document}
