+++ server/resources/preamble.tex
@@ -1,0 +1,203 @@
+\IfFileExists{stacks-project.cls}{%
+\documentclass{stacks-project}
+}{%
+\documentclass{amsart}
+}
+
+% For dealing with references we use the comment environment
+\usepackage{verbatim}
+\newenvironment{reference}{\comment}{\endcomment}
+%\newenvironment{reference}{}{}
+\newenvironment{slogan}{\comment}{\endcomment}
+\newenvironment{history}{\comment}{\endcomment}
+
+% For commutative diagrams we use Xy-pic
+\usepackage[all]{xy}
+
+% We use 2cell for 2-commutative diagrams.
+\xyoption{2cell}
+\UseAllTwocells
+
+% We use multicol for the list of chapters between chapters
+\usepackage{multicol}
+
+% This is generall recommended for better output
+\usepackage[T1]{fontenc}
+
+% For cross-file-references
+\usepackage{xr-hyper}
+
+% Package for hypertext links:
+\usepackage{hyperref}
+
+% For any local file, say "hello.tex" you want to link to please
+% use \externaldocument[hello-]{hello}
+\externaldocument[introduction-]{introduction}
+\externaldocument[conventions-]{conventions}
+\externaldocument[sets-]{sets}
+\externaldocument[categories-]{categories}
+\externaldocument[topology-]{topology}
+\externaldocument[sheaves-]{sheaves}
+\externaldocument[sites-]{sites}
+\externaldocument[stacks-]{stacks}
+\externaldocument[fields-]{fields}
+\externaldocument[algebra-]{algebra}
+\externaldocument[brauer-]{brauer}
+\externaldocument[homology-]{homology}
+\externaldocument[derived-]{derived}
+\externaldocument[simplicial-]{simplicial}
+\externaldocument[more-algebra-]{more-algebra}
+\externaldocument[smoothing-]{smoothing}
+\externaldocument[modules-]{modules}
+\externaldocument[sites-modules-]{sites-modules}
+\externaldocument[injectives-]{injectives}
+\externaldocument[cohomology-]{cohomology}
+\externaldocument[sites-cohomology-]{sites-cohomology}
+\externaldocument[dga-]{dga}
+\externaldocument[dpa-]{dpa}
+\externaldocument[hypercovering-]{hypercovering}
+\externaldocument[schemes-]{schemes}
+\externaldocument[constructions-]{constructions}
+\externaldocument[properties-]{properties}
+\externaldocument[morphisms-]{morphisms}
+\externaldocument[coherent-]{coherent}
+\externaldocument[divisors-]{divisors}
+\externaldocument[limits-]{limits}
+\externaldocument[varieties-]{varieties}
+\externaldocument[topologies-]{topologies}
+\externaldocument[descent-]{descent}
+\externaldocument[perfect-]{perfect}
+\externaldocument[more-morphisms-]{more-morphisms}
+\externaldocument[flat-]{flat}
+\externaldocument[groupoids-]{groupoids}
+\externaldocument[more-groupoids-]{more-groupoids}
+\externaldocument[etale-]{etale}
+\externaldocument[chow-]{chow}
+\externaldocument[intersection-]{intersection}
+\externaldocument[pic-]{pic}
+\externaldocument[adequate-]{adequate}
+\externaldocument[dualizing-]{dualizing}
+\externaldocument[duality-]{duality}
+\externaldocument[discriminant-]{discriminant}
+\externaldocument[local-cohomology-]{local-cohomology}
+\externaldocument[algebraization-]{algebraization}
+\externaldocument[curves-]{curves}
+\externaldocument[resolve-]{resolve}
+\externaldocument[models-]{models}
+\externaldocument[pione-]{pione}
+\externaldocument[etale-cohomology-]{etale-cohomology}
+\externaldocument[proetale-]{proetale}
+\externaldocument[crystalline-]{crystalline}
+\externaldocument[spaces-]{spaces}
+\externaldocument[spaces-properties-]{spaces-properties}
+\externaldocument[spaces-morphisms-]{spaces-morphisms}
+\externaldocument[decent-spaces-]{decent-spaces}
+\externaldocument[spaces-cohomology-]{spaces-cohomology}
+\externaldocument[spaces-limits-]{spaces-limits}
+\externaldocument[spaces-divisors-]{spaces-divisors}
+\externaldocument[spaces-over-fields-]{spaces-over-fields}
+\externaldocument[spaces-topologies-]{spaces-topologies}
+\externaldocument[spaces-descent-]{spaces-descent}
+\externaldocument[spaces-perfect-]{spaces-perfect}
+\externaldocument[spaces-more-morphisms-]{spaces-more-morphisms}
+\externaldocument[spaces-flat-]{spaces-flat}
+\externaldocument[spaces-groupoids-]{spaces-groupoids}
+\externaldocument[spaces-more-groupoids-]{spaces-more-groupoids}
+\externaldocument[bootstrap-]{bootstrap}
+\externaldocument[spaces-pushouts-]{spaces-pushouts}
+\externaldocument[spaces-chow-]{spaces-chow}
+\externaldocument[groupoids-quotients-]{groupoids-quotients}
+\externaldocument[spaces-more-cohomology-]{spaces-more-cohomology}
+\externaldocument[spaces-simplicial-]{spaces-simplicial}
+\externaldocument[spaces-duality-]{spaces-duality}
+\externaldocument[formal-spaces-]{formal-spaces}
+\externaldocument[restricted-]{restricted}
+\externaldocument[spaces-resolve-]{spaces-resolve}
+\externaldocument[formal-defos-]{formal-defos}
+\externaldocument[defos-]{defos}
+\externaldocument[cotangent-]{cotangent}
+\externaldocument[examples-defos-]{examples-defos}
+\externaldocument[algebraic-]{algebraic}
+\externaldocument[examples-stacks-]{examples-stacks}
+\externaldocument[stacks-sheaves-]{stacks-sheaves}
+\externaldocument[criteria-]{criteria}
+\externaldocument[artin-]{artin}
+\externaldocument[quot-]{quot}
+\externaldocument[stacks-properties-]{stacks-properties}
+\externaldocument[stacks-morphisms-]{stacks-morphisms}
+\externaldocument[stacks-limits-]{stacks-limits}
+\externaldocument[stacks-cohomology-]{stacks-cohomology}
+\externaldocument[stacks-perfect-]{stacks-perfect}
+\externaldocument[stacks-introduction-]{stacks-introduction}
+\externaldocument[stacks-more-morphisms-]{stacks-more-morphisms}
+\externaldocument[stacks-geometry-]{stacks-geometry}
+\externaldocument[moduli-]{moduli}
+\externaldocument[moduli-curves-]{moduli-curves}
+\externaldocument[examples-]{examples}
+\externaldocument[exercises-]{exercises}
+\externaldocument[guide-]{guide}
+\externaldocument[desirables-]{desirables}
+\externaldocument[coding-]{coding}
+\externaldocument[obsolete-]{obsolete}
+\externaldocument[fdl-]{fdl}
+\externaldocument[index-]{index}
+
+% Theorem environments.
+%
+\theoremstyle{plain}
+\newtheorem{theorem}[subsection]{Theorem}
+\newtheorem{proposition}[subsection]{Proposition}
+\newtheorem{lemma}[subsection]{Lemma}
+
+\theoremstyle{definition}
+\newtheorem{definition}[subsection]{Definition}
+\newtheorem{example}[subsection]{Example}
+\newtheorem{exercise}[subsection]{Exercise}
+\newtheorem{situation}[subsection]{Situation}
+
+\theoremstyle{remark}
+\newtheorem{remark}[subsection]{Remark}
+\newtheorem{remarks}[subsection]{Remarks}
+
+\numberwithin{equation}{subsection}
+
+% Macros
+%
+\def\lim{\mathop{\mathrm{lim}}\nolimits}
+\def\colim{\mathop{\mathrm{colim}}\nolimits}
+\def\Spec{\mathop{\mathrm{Spec}}}
+\def\Hom{\mathop{\mathrm{Hom}}\nolimits}
+\def\Ext{\mathop{\mathrm{Ext}}\nolimits}
+\def\SheafHom{\mathop{\mathcal{H}\!\mathit{om}}\nolimits}
+\def\SheafExt{\mathop{\mathcal{E}\!\mathit{xt}}\nolimits}
+\def\Sch{\mathit{Sch}}
+\def\Mor{\mathop{Mor}\nolimits}
+\def\Ob{\mathop{\mathrm{Ob}}\nolimits}
+\def\Sh{\mathop{\mathit{Sh}}\nolimits}
+\def\NL{\mathop{N\!L}\nolimits}
+\def\proetale{{pro\text{-}\acute{e}tale}}
+\def\etale{{\acute{e}tale}}
+\def\QCoh{\mathit{QCoh}}
+\def\Ker{\mathop{\mathrm{Ker}}}
+\def\Im{\mathop{\mathrm{Im}}}
+\def\Coker{\mathop{\mathrm{Coker}}}
+\def\Coim{\mathop{\mathrm{Coim}}}
+
+%
+% Macros for moduli stacks/spaces
+%
+\def\QCohstack{\mathcal{QC}\!\mathit{oh}}
+\def\Cohstack{\mathcal{C}\!\mathit{oh}}
+\def\Spacesstack{\mathcal{S}\!\mathit{paces}}
+\def\Quotfunctor{\mathrm{Quot}}
+\def\Hilbfunctor{\mathrm{Hilb}}
+\def\Curvesstack{\mathcal{C}\!\mathit{urves}}
+\def\Polarizedstack{\mathcal{P}\!\mathit{olarized}}
+\def\Complexesstack{\mathcal{C}\!\mathit{omplexes}}
+% \Pic is the operator that assigns to X its picard group, usage \Pic(X)
+% \Picardstack_{X/B} denotes the Picard stack of X over B
+% \Picardfunctor_{X/B} denotes the Picard functor of X over B
+\def\Pic{\mathop{\mathrm{Pic}}\nolimits}
+\def\Picardstack{\mathcal{P}\!\mathit{ic}}
+\def\Picardfunctor{\mathrm{Pic}}
+\def\Deformationcategory{\mathcal{D}\!\mathit{ef}}
